+++ lomap/tests/README.md
@@ -1,0 +1,10 @@
+File Descriptions:
+* `test_map_word_accepted_randomized_occupancy_grid.py`. Creates a FSA using LOMAP. Creates a randomized occupancy grid with a randomized start and goal location. The occupancy grid uses values -1 for occupied, 0 for open, and a binary value relating to symbols of the alphabet from the FSA. The parameter `percent_empty_open` determines the percent of the total occupancy grid spaces that have a value of 0. The remaining spaces are either -1 or the binary values. A* is run to find a path between the randomized start and goal location using the occupanchy grid, and the custom A* produces a path and the symbols produced by the occupancy grid along that path. This path of symbols is transformed from their numerical binary represnetation into the alphetical representation used by the FSA. This path aka word is run through the FSA's `is_word_accepted_verbose` method. If true, the word produced by A* is accepted by the FSA, if it is false the word is not.
+
+* `test_map_word_accepted.py`. Creates a FSA using LOMAP. Creates an occupancy grid from a specified csv. The csv contains the grid along with a start and goal location. The occupancy grid uses values -1 for occupied, 0 for open, and a binary value relating to symbols of the alphabet from the FSA. A* is run to find a path between the randomized start and goal location using the occupanchy grid, and the custom A* produces a path and the symbols produced by the occupancy grid along that path. This path of symbols is transformed from their numerical binary represnetation into the alphetical representation used by the FSA. This path aka word is run through the FSA's `is_word_accepted_verbose` method. If true, the word produced by A* is accepted by the FSA, if it is false the word is not.
+
+* `test_tl_composition_graph_traversal`. Uses logic from `test_map_word_accepted.py` and `test_map_word_accepted_randomized_occupancy_grid.py` to create an FSA from a spec using LOMAP. Creates a transition graph from an occupancy grid representation. Work in progress.
+
+* `map.csv`. Map containing spacecs with single symbols, e.g. A or B.
+
+* `map_multiple_symbols.csv`. Map containing spaces with single and multiple symbols e.g. A, B, A&B.
